+++ supreme/common/src/commonMain/kotlin/KeyConstraints.kt
@@ -1,11 +1,8 @@
 package at.asitplus.attestation.supreme
 
-import at.asitplus.catchingUnwrapped
-import at.asitplus.signum.indispensable.CryptoPublicKey
 import at.asitplus.signum.indispensable.Digest
 import at.asitplus.signum.indispensable.ECCurve
 import at.asitplus.signum.indispensable.RSAPadding
-import at.asitplus.signum.indispensable.SignatureAlgorithm
 import at.asitplus.signum.indispensable.misc.BitLength
 import at.asitplus.signum.indispensable.nativeDigest
 import kotlinx.serialization.SerialName
@@ -42,9 +39,10 @@ data class KeyConstraints(
             val keySize: @Serializable(with = BitLengthSerializer::class) BitLength,
             val paddings: Set<RSAPadding> = setOf(RSAPadding.PSS),
             override val digests: Set<Digest> = setOf(Digest.SHA256),
-            override val allowSigning: Boolean = true,
             val allowDecrypting: Boolean = false,
         ) : AlgorithmParameters() {
+            //must be true as of now, because we require signing for proof of possession
+            override val allowSigning: Boolean = true
             override fun equals(other: Any?): Boolean {
                 if (this === other) return true
                 if (other !is RSA) return false
@@ -76,9 +74,10 @@ data class KeyConstraints(
             @Serializable(with = ECCurveSerializer::class)
             val curve: ECCurve = ECCurve.SECP_256_R_1,
             override val digests: Set<@Serializable(with = DigestSerializer::class) Digest> = setOf(curve.nativeDigest),
-            override val allowSigning: Boolean = true,
             val allowKeyAgreement: Boolean = false,
         ) : AlgorithmParameters() {
+            //must be true as of now, because we require signing for proof of possession
+            override val allowSigning: Boolean = true
             override fun equals(other: Any?): Boolean {
                 if (this === other) return true
                 if (other !is EC) return false
@@ -120,7 +119,7 @@ data class KeyConstraints(
     }
 
     @Serializable
-    class KeyProtection(
+    data class KeyProtection(
         val timeout: Duration? = null,
         val deviceLock: Boolean? = null,
         val biometry: Boolean? = null,
@@ -163,4 +162,6 @@ data class KeyConstraints(
         return result
     }
 
+    @Serializable
+    data class AuthPrompt(val message: String, val cancelText: String)
 }