+++ supreme/common/src/commonTest/kotlin/alibiTest.kt
@@ -1,0 +1,6 @@
+import at.asitplus.testballoon.invoke
+import de.infix.testBalloon.framework.core.testSuite
+
+val alibiTest by testSuite {
+    "test are running" {}
+}