+++ supreme/verifier/src/commonTest/kotlin/examples/Readme-Verifier.kt
@@ -16,6 +16,11 @@ import kotlin.time.Duration.Companion.seconds
 val redisBacked: ChallengeValidator = TODO()
 val serviceSpecificOID = WardenDefaults.OIDs.ATTESTATION_PROOF
 
+
+
+
+
+
 val verifier = AttestationVerifier(
     makoto = makoto,
  /*(1)!*/attestationProofOID = serviceSpecificOID, //override default
@@ -24,14 +29,13 @@ val verifier = AttestationVerifier(
         algorithmParameters = AlgorithmParameters.EC(
             curve = ECCurve.SECP_256_R_1,
             digests = setOf(ECCurve.SECP_256_R_1.nativeDigest),
-            allowSigning = true, //DEFAULT
             allowKeyAgreement = false //DEFAULT
         ),
  /*(4)!*/keyProtection = KeyProtection(
             timeout = 30.seconds,
             deviceLock = false,
             biometry = true,
-            allowNewBiometricFactors = false
+            allowNewBiometricFactors = false,
         )
     ),
     nonceValidity = 5.minutes, //DEFAULT